# docs/api/openapi-spec.yml
...
openapi: 3.1.0
info:
  title: LogPilot Agent API
  description: 'Cloud-native, autonomous log-intelligence agent: ingestion, search, chat, RCA, reports, proactive
    failure forecasting and autonomy policy.


    Authenticate with POST /api/v1/auth/login, then send `Authorization: Bearer <token>`. Real-time events and streamed
    chat use the WebSocket at /ws?token=<jwt> (not described by OpenAPI).'
  version: 2.0.0
paths:
...
  /api/v1/policy/tools/{tool_name}:
    put:
      tags:
      - policy
      summary: Set Policy
      operationId: set_policy_api_v1_policy_tools__tool_name__put
      security:
      - HTTPBearer: []
      parameters:
      - name: tool_name
        in: path
        required: true
        schema:
          type: string
          title: Tool Name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - policy
      summary: Delete Override
      description: Remove an environment-specific override (the '*' row is the base policy and stays).
      operationId: delete_override_api_v1_policy_tools__tool_name__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: tool_name
        in: path
        required: true
        schema:
          type: string
          title: Tool Name
      - name: scope
        in: query
        required: true
        schema:
          type: string
          title: Scope
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PolicyBody:
      properties:
        scope:
          type: string
          title: Scope
          default: '*'
        tier:
          type: string
          enum:
          - read_only
          - autonomous_background
          - autonomous_policy_bounded
          - propose_only
          - autonomous_execution
          title: Tier
        min_confidence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Min Confidence
          default: 0.0
        requires_approval:
          type: boolean
          title: Requires Approval
          default: false
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - tier
      title: PolicyBody
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://localhost:8000
  description: Local (docker compose)